djanmongo/requirements.txt
Django>=4.0
djangorestframework>=3.13
djangorestframework-simplejwt>=5.0
psycopg2-binary>=2.9 # Or your preferred DB driver (mysqlclient, etc.)
# djangocorsheaders # Often needed for frontend interaction
django-unfold-0.54.0